LAB02/03-CloudAlbum-COGNITO/requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.1.0
Pillow==5.1.0
PyMySQL==0.8.1
requests==2.19.1
requests-oauthlib==1.0.0
SQLAlchemy==1.2.8
urllib3==1.23
uWSGI==2.0.17
Werkzeug==0.14.1
WTForms==2.2.1
import_string
pynamodb
python_jose<2.1